Optimize qk rope/nope concat
